// server/src/types/model.ts
export interface AddressIds {
  startAddressId: number;
  destinationAddressId: number;
}

export interface CityIds {
  startCityId: number;
  destinationCityId: number;
}

export interface CountryIds {
  startCountryId: number;
  destinationCountryId: number;
}

export interface StreetIds {
  startStreetId: number;
  destinationStreetId: number;
}